🚚 move setups to pageSetups
