+++ test/e2e/lib/testSetup/index.ts
@@ -1,4 +1,4 @@
 export { createTest } from './createTest'
-export { bundleSetup, html } from './setups'
+export { bundleSetup, html } from './pageSetups'
 export { EventRegistry } from './eventsRegistry'
 export { getTestServers } from '../servers'
